Se agrega la busqueda de usuarios (falta despues los ajustes cuando este terminado el backend). Se apunta al nuevo servidor de heroku.

+++ app/src/main/java/com/example/marco/fiubados/commons/FieldsValidator.java
@@ -33,7 +33,7 @@ public class FieldsValidator {
      */
     public static boolean isNumericFieldValid(String fieldValue){
         try {
-            Integer.parseInt(fieldValue);
+            Long.parseLong(fieldValue);
         } catch (NumberFormatException nfe){
             return false;
         }
